pyapp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pymongo==4.8.0
PyJWT==2.9.0
bcrypt==4.1.3
python-multipart==0.0.9